// mobile/enums/UsuarioPrefs.ts
export enum PrefIdioma {
    Portugues = 'portugues',
    Ingles = 'ingles',
    Espanhol = 'espanhol',
    Frances = 'frances',
}

export enum PrefTransporte {
    Carro = 'carro',
    Onibus = 'onibus',
    Aviao = 'aviao',
    Trem = 'trem',
}

export enum PrefAcomodacao {
    Hotel = 'hotel',
    Apartamento = 'apartamento',
    Hostel = 'hostel',
    Pousada = 'pousada',
}

export enum PrefOrcamento {
    Baixo = 'baixo',
    Medio = 'medio',
    Alto = 'alto',
}